refac: move type `DhatuDetails` to 'utils/types'
Why?
Required by two files - but cannot be colocated with neither -
since both are equally related.
