+++ src/app/dhatu/[id]/DetailsTable.tsx
@@ -6,7 +6,7 @@ import Col from "react-bootstrap/Col";
 import Table from "react-bootstrap/Table";
 import Card from "react-bootstrap/Card";
 
-import { DhatuDetails } from "@/utils/getDhatupathaLocal";
+import type { DhatuDetails } from "@/utils/types";
 
 import styles from "./styles.module.scss";
