+++ src/utils/getDhatupathaLocal.ts
@@ -1,20 +1,6 @@
 import fs from "fs/promises";
 import path from "path";
-
-export type DhatuDetails = {
-  id: number;
-  dhatuId: string;
-  muladhatu: string;
-  dhatu: string;
-  graphURL: string;
-  meaning: string;
-  gana: string;
-  padi: string;
-  it: string;
-  madhaviyaId: string;
-  kshirataranginiId: string;
-  dhatupradipaId: string;
-};
+import { DhatuDetails } from "./types";
 
 const dhatupathaPath = "src/assets/dhatupatha.json";
