+++ src/utils/types.ts
@@ -1,0 +1,14 @@
+export type DhatuDetails = {
+  id: number;
+  dhatuId: string;
+  muladhatu: string;
+  dhatu: string;
+  graphURL: string;
+  meaning: string;
+  gana: string;
+  padi: string;
+  it: string;
+  madhaviyaId: string;
+  kshirataranginiId: string;
+  dhatupradipaId: string;
+};
